need an interface for the oracle table name getter since we can't mock final
